+++ src/services/upload.ts
@@ -1,10 +1,10 @@
-import axios from 'axios';
-import baseUrl from './api';
+import axios from "axios";
+import baseUrl from "./api";
 
-const PostFileUpload = (formData) => {
+const PostFileUpload = (formData: any) => {
   const config = {
     headers: {
-      'Content-Type': 'multipart/form-data',
+      "Content-Type": "multipart/form-data",
     },
   };
   return axios.post(`${baseUrl}/v1/file/upload`, formData, config);
